requirements.txt
Brotli==1.1.0
MarkupSafe==3.0.2
PySide6==6.8.2
PySocks==1.7.1
PyYAML==6.0.2
Send2Trash==1.8.3
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autocommand==2.2.2
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
cached-property==1.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
comm==0.2.2
conda-pack==0.0.0
contourpy==1.3.1
cryptography==46.0.6
cuda-bindings==13.1.1
cuda-core==0.6.0
cuda-pathfinder==1.4.0
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fonttools==4.56.0
fqdn==1.5.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.10
ihm==2.3
importlib-metadata==8.6.1
importlib-resources==6.5.2
inflect==7.3.1
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.15.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
llvmlite==0.44.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.26.0
mistune==3.1.3
more-itertools==10.3.0
msgpack==1.1.0
munkres==1.1.4
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.3
notebook-shim==0.2.4
numba==0.61.0
numba-cuda==0.28.2
numpy==2.1.3
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
opencv-python==4.11.0
opencv-python-headless==4.11.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-stubs==2.3.0.250703
pandocfilters==1.5.0
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pip==25.0.1
pkgutil-resolve-name==1.3.10
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.6.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.19.1
pyjwt==2.12.1
pyparsing==3.2.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==2.0.7
python-multipart==0.0.22
pytz==2024.1
pywin32==311
pywinpty==2.0.15
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.23.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==75.8.2
shiboken6==6.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
sse-starlette==3.3.4
stack-data==0.6.3
starlette==1.0.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.3.0
types-python-dateutil==2.9.0.20241206
typing-extensions==4.15.0
typing-inspection==0.4.2
typing-utils==0.1.0
tzdata==2025.2
unicodedata2==16.0.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.42.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.13
win-inet-pton==1.1.0
zipp==3.21.0
zstandard==0.23.0
